+++ KlaviyoV3Sdk/KlaviyoV3Api.php
@@ -25,7 +25,8 @@ class KlaviyoV3Api
     /**
      * Error messages
      */
-    const ERROR_INVALID_API_KEY = 'The Private Klaviyo API Key you have set is invalid.';
+    const ERROR_FORBIDDEN = 'Private API key has invalid permissions.';
+    const ERROR_NOT_AUTHORIZED = 'Authentication key missing from request or is invalid. Check that Klaviyo Private API key is correctly set for scope.';
     const ERROR_NON_200_STATUS = 'Request Failed with HTTP Status Code: %s';
     const ERROR_API_CALL_FAILED = 'Request could be completed at this time, API call failed';
     const ERROR_MALFORMED_RESPONSE_BODY = 'Response from API could not be decoded from JSON, check response body';
@@ -190,8 +191,10 @@ class KlaviyoV3Api
     {
         $body = array(
             self::DATA_KEY_PAYLOAD => array(
-                self::TYPE_KEY_PAYLOAD => self::PROFILE_KEY_PAYLOAD,
-                self::ID_KEY_PAYLOAD => $profile_id
+                array(
+                    self::TYPE_KEY_PAYLOAD => self::PROFILE_KEY_PAYLOAD,
+                    self::ID_KEY_PAYLOAD => $profile_id
+                )
             )
         );
 
@@ -209,14 +212,16 @@ class KlaviyoV3Api
     public function createProfile($profile_properties)
     {
         $body = array(
-            self::DATA_KEY_PAYLOAD => array(
-                self::TYPE_KEY_PAYLOAD => self::PROFILE_KEY_PAYLOAD,
-                self::ATTRIBUTE_KEY_PAYLOAD => $profile_properties
-            )
+            self::DATA_KEY_PAYLOAD =>
+                array(
+                    self::TYPE_KEY_PAYLOAD => self::PROFILE_KEY_PAYLOAD,
+                    self::ATTRIBUTE_KEY_PAYLOAD => $profile_properties
+                )
+
         );
 
         $response_body = $this->requestV3('api/profiles/', self::HTTP_POST, $body);
-        $id = $response_body[self::DATA_KEY_PAYLOAD][0][self::ID_KEY_PAYLOAD];
+        $id = $response_body[self::DATA_KEY_PAYLOAD][self::ID_KEY_PAYLOAD];
         return [
             'data' => $response_body,
             'profile_id' => $id
@@ -364,7 +369,7 @@ class KlaviyoV3Api
                 sleep(1);
                 $this->requestV3($path, $method, $body, $attempt + 1);
             } else {
-                throw new KlaviyoApiException(self::ERROR_API_CALL_FAILED);
+                $this->handleAPIResponse($response, $statusCode);
             }
         }
         curl_close($curl);
@@ -463,14 +468,16 @@ class KlaviyoV3Api
     protected function handleAPIResponse($response, $statusCode)
     {
         $decoded_response = $this->decodeJsonResponse($response);
-        if ($statusCode == 403) {
-            throw new KlaviyoAuthenticationException(self::ERROR_INVALID_API_KEY, $statusCode);
+        if ($statusCode == 401) {
+            throw new KlaviyoAuthenticationException(self::ERROR_NOT_AUTHORIZED, $statusCode);
+        } elseif ($statusCode == 403) {
+            throw new KlaviyoAuthenticationException(self::ERROR_FORBIDDEN, $statusCode);
         } elseif ($statusCode == 429) {
             throw new KlaviyoRateLimitException(
                 self::ERROR_RATE_LIMIT_EXCEEDED
             );
         } elseif ($statusCode < 200 || $statusCode >= 300) {
-            throw new KlaviyoApiException(isset($decoded_response['detail']) ? $decoded_response['detail'] : sprintf(self::ERROR_NON_200_STATUS, $statusCode), $statusCode);
+            throw new KlaviyoApiException(isset($decoded_response['errors']) ? $decoded_response['errors'][0]['detail'] : sprintf(self::ERROR_NON_200_STATUS, $statusCode), $statusCode);
         }
 
         return $decoded_response;